AI mini projects/Fashion-MNIST-recognizer/requirements.txt
numpy
matplotlib
opencv-python
torch
torchvision
easydict
pillow
python-dotenv
mlflow